+++ src/main/java/domain/Dealer.java
@@ -1,34 +1,14 @@
 package domain;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
-public class Dealer {
+public class Dealer extends Participant{
 
-    private final List<Card> hand = new ArrayList<>();
-
-    private Map<GameResult, Integer> gameResults = new HashMap<>();
+    private final Map<GameResult, Integer> gameResults = new HashMap<>();
 
     public Card reveal() {
         return hand.getFirst();
-    }
-
-    public void receiveInitCard(List<Card> cards) {
-        hand.addAll(cards);
-    }
-
-    public void receiveCard(Card card) {
-        hand.add(card);
-    }
-
-    public List<Card> getHand() {
-        return hand;
-    }
-
-    public void setRounds(GameResult gameResult) {
-        gameResults.put(gameResult, gameResults.getOrDefault(gameResult,0) + 1);
     }
 
     public int getWinRounds() {
@@ -41,5 +21,9 @@ public class Dealer {
 
     public int getLoseRounds() {
         return gameResults.getOrDefault(GameResult.LOSE,0);
+    }
+
+    public void setRounds(GameResult gameResult) {
+        gameResults.put(gameResult, gameResults.getOrDefault(gameResult,0) + 1);
     }
 }
